fix issues with directory mapping
the buildSourceTree script creates the $SubD based off of the location of the script. However when it creates the pullRepo script it uses "~" which is the root of the user's folder and not the location of the script. Adjusting to be based on the location of the buildSourceTree script makes this work correctly